+++ task_core/db_publish.py
@@ -26,7 +26,6 @@ from task_core.cleanup import attempt_all_cleanup
 from task_core.db_insert import load_rows_into_staging
 from task_core.db_copy import (
     SpoolIdentity,
-    cleanup_default_spool_directory,
     cleanup_predecessor_spools,
     cleanup_spool_paths,
     load_copy_into_staging,
@@ -1444,7 +1443,6 @@ class DbPublisher:
                 self.task_name,
                 ', '.join(str(path) for path in deleted),
             )
-        cleanup_default_spool_directory(self.copy_load_policy)
         return deleted
 
     def _require_task_lock(self, action):
@@ -1823,7 +1821,6 @@ class DbPublisher:
                         f'could not remove completed COPY spool(s): '
                         f'{failed_cleanup!r}'
                     )
-                cleanup_default_spool_directory(self.copy_load_policy)
                 prepared_copy = None
 
             self._verify_prepared_table(
@@ -1854,7 +1851,6 @@ class DbPublisher:
         except BaseException:
             if prepared_copy is not None:
                 cleanup_spool_paths([prepared_copy.path])
-                cleanup_default_spool_directory(self.copy_load_policy)
             raise
 
     def _verify_prepared_table(